refactor: after code review

+++ src/main/java/com/cool/movie/service/impl/MovieServiceImpl.java
@@ -49,7 +49,7 @@ public class MovieServiceImpl implements MovieService {
                 , pageNumber
                 , moviesFind.getTotalPages()
                 , (int) moviesFind.getTotalElements()
-                , (int) moviesFind.getNumberOfElements()
+                , moviesFind.getNumberOfElements()
                 , movieMapper.toResponses(moviesFind.toList()));
 
         return moviePage;
